tools/luban-provisioner/templates/source/luban-dagster-dbt-starrocks-code-location-source-template/{{cookiecutter.app_name}}/src/{{cookiecutter.package_name}}/resources/starrocks.py
from __future__ import annotations

import os
from dataclasses import dataclass
from typing import Any


@dataclass(frozen=True)
class StarRocksClient:
    host: str
    port: int
    user: str
    password: str
    connect_timeout: int = 10

    def query_scalar(self, sql: str) -> Any:
        import pymysql

        connection = pymysql.connect(
            host=self.host,
            port=self.port,
            user=self.user,
            password=self.password,
            connect_timeout=self.connect_timeout,
            read_timeout=self.connect_timeout,
            write_timeout=self.connect_timeout,
            autocommit=True,
        )
        try:
            with connection.cursor() as cursor:
                cursor.execute(sql)
                row = cursor.fetchone()
                return row[0] if row else None
        finally:
            connection.close()

    def query_first_column(self, sql: str) -> list[Any]:
        import pymysql

        connection = pymysql.connect(
            host=self.host,
            port=self.port,
            user=self.user,
            password=self.password,
            connect_timeout=self.connect_timeout,
            read_timeout=self.connect_timeout,
            write_timeout=self.connect_timeout,
            autocommit=True,
        )
        try:
            with connection.cursor() as cursor:
                cursor.execute(sql)
                rows = cursor.fetchall()
                return [row[0] for row in rows]
        finally:
            connection.close()


def make_starrocks_resource() -> StarRocksClient:
    return StarRocksClient(
        host=os.getenv("STARROCKS_HOST", "localhost"),
        port=int(os.getenv("STARROCKS_PORT", "9030")),
        user=os.getenv("STARROCKS_USER", "root"),
        password=os.getenv("STARROCKS_PASSWORD", ""),
    )

